packages/tools/mcp_servers/pyproject.toml
[project]
name = "mcp-servers"
version = "0.0.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = ["fastapi", "langchain-mcp-adapters"]
